moved all payment types and interfaces to payment and extended subd0cument interfaces with Document
