+++ src/lib/models/Payment.ts
@@ -1,8 +1,26 @@
 import { Schema, model, ObjectId, Document, models } from "mongoose"
-import { IPaymentDetails } from "./User"
 
-// Custom interfaces for Payment Schema
-export interface IPayment extends IPaymentDetails {
+// Custom types and interfaces for Payment Schema
+export type IPaymentType =
+	| "cash"
+	| "credit"
+	| "debit"
+	| "bank"
+	| "money order"
+	| "check"
+
+export interface IPaymentDetails {
+	method: IPaymentType
+	cardHolderName?: string
+	cardNumber?: string
+	expirationDate?: Date
+	cvv?: string
+	routingNumber?: string
+	accountNumber?: string
+	bankName?: string
+}
+
+export interface IPayment extends IPaymentDetails, Document {
 	_id: ObjectId
 	bookingId: ObjectId // reference to the booking the payment is associated with
 	amount: number
@@ -10,6 +28,8 @@ export interface IPayment extends IPaymentDetails {
 	createdAt: Date
 	createdBy: string // reference to the user (employee) who created the payment - uses uuid v4
 }
+
+// TODO: Add validation, error messages, and hooks to the Payment Schema
 
 // Payment Schema - subdocument of Booking Schema
 const paymentSchema = new Schema<IPayment>(
