+++ src/lib/models/index.ts
@@ -1,4 +1,5 @@
 import { User } from "./User"
 import { Booking } from "./Booking"
+import { Payment } from "./Payment"
 
-export { User, Booking }
+export { User, Booking, Payment }
